moved definitions [run CI]
