+++ BodoSQL/bodosql/py4j_gateway.py
@@ -5,7 +5,7 @@ APIs used to launch and connect to the Java py4j calcite gateway server.
 import os
 import sys
 import warnings
-from typing import Any, cast
+from typing import cast
 
 from py4j.java_gateway import GatewayParameters, JavaGateway, launch_gateway
 
@@ -117,33 +117,3 @@ def configure_java_logging(level: int):
         from bodosql.imported_java_classes import JavaEntryPoint
 
         JavaEntryPoint.configureJavaLogging(level)
-
-
-def build_java_array_list(elems: list[Any]):
-    if bodo.get_rank() == 0:
-        from bodosql.imported_java_classes import JavaEntryPoint
-
-        output_list = JavaEntryPoint.buildArrayList()
-        for elem in elems:
-            JavaEntryPoint.appendToArrayList(output_list, elem)
-        return output_list
-
-
-def build_java_hash_map(d: dict[Any, Any]):
-    if bodo.get_rank() == 0:
-        from bodosql.imported_java_classes import JavaEntryPoint
-
-        output_map = JavaEntryPoint.buildMap()
-        for key, value in d.items():
-            JavaEntryPoint.mapPut(output_map, key, value)
-        return output_map
-
-
-def build_java_properties(d: dict[str, str]):
-    if bodo.get_rank() == 0:
-        from bodosql.imported_java_classes import JavaEntryPoint
-
-        output_map = JavaEntryPoint.buildProperties()
-        for key, value in d.items():
-            JavaEntryPoint.setProperty(output_map, key, value)
-        return output_map
